+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Validate image url provided as external url for user profile picture. Refs UIU-3080.
 * Import and use ProfilePicture Component from stripes/smart-components. Refs UIU-3104.
 * Show Roles assigned to users. Refs UIU-3110.
+* Add edit user roles accordion on edit role modal view. Refs UIU-3021.    
 
 ## [10.1.0](https://github.com/folio-org/ui-users/tree/v10.1.0) (2024-03-20)
 [Full Changelog](https://github.com/folio-org/ui-users/compare/v10.0.4...v10.1.0)